# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YOGA LIFE VN API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new student/member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /centers:
    get:
      operationId: listCenters
      tags: [centers]
      summary: List all yoga centers/branches
      responses:
        "200":
          description: List of centers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CentersListResponse"
    post:
      operationId: createCenter
      tags: [centers]
      summary: Create a new center/branch (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCenterRequest"
      responses:
        "201":
          description: Created center
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Center"
  /centers/{id}:
    get:
      operationId: getCenter
      tags: [centers]
      summary: Get center by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Center details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Center"
    put:
      operationId: updateCenter
      tags: [centers]
      summary: Update center
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCenterRequest"
      responses:
        "200":
          description: Updated center
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Center"
    delete:
      operationId: deleteCenter
      tags: [centers]
      summary: Delete center
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /centers/{id}/assign-admin:
    post:
      operationId: assignCenterAdmin
      tags: [centers]
      summary: Assign admin to center (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAdminRequest"
      responses:
        "200":
          description: Admin assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      parameters:
        - name: centerId
          in: query
          schema:
            type: integer
        - name: instructorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassesListResponse"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YogaClass"
  /classes/{id}:
    put:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassRequest"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YogaClass"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        phone:
          type: string
      required: [email, password, fullName]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, center_admin, super_admin]
        centerId:
          type: integer
          nullable: true
        centerName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, fullName, role, createdAt]
...
    Center:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        adminId:
          type: integer
          nullable: true
        adminName:
          type: string
          nullable: true
        memberCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, address, city, phone, email, memberCount, createdAt]
    CreateCenterRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: [name, address, city, phone, email]
    UpdateCenterRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
    AssignAdminRequest:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    CentersListResponse:
      type: object
      properties:
        centers:
          type: array
          items:
            $ref: "#/components/schemas/Center"
        total:
          type: integer
      required: [centers, total]
...
    YogaClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        instructorId:
          type: integer
          nullable: true
        instructorName:
          type: string
          nullable: true
        centerId:
          type: integer
        centerName:
          type: string
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        level:
          type: string
          enum: [beginner, intermediate, advanced, all_levels]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, centerId, centerName, dayOfWeek, startTime, endTime, capacity, level, isActive, createdAt]
    CreateClassRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        instructorId:
          type: integer
          nullable: true
        centerId:
          type: integer
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        level:
          type: string
          enum: [beginner, intermediate, advanced, all_levels]
        isActive:
          type: boolean
      required: [name, centerId, dayOfWeek, startTime, endTime, capacity, level]
    UpdateClassRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        instructorId:
          type: integer
          nullable: true
        dayOfWeek:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        capacity:
          type: integer
        level:
          type: string
          enum: [beginner, intermediate, advanced, all_levels]
        isActive:
          type: boolean
    ClassesListResponse:
      type: object
      properties:
        classes:
          type: array
          items:
            $ref: "#/components/schemas/YogaClass"
        total:
          type: integer
      required: [classes, total]
...